resolves #2: added custom kernel recipe (#16)
* resolves #1: added own layer

* added vscode workspace configurations

This can be use with Yocto's vscode extension

* resolves #2: added kernel recipe

This recipe cannot build yet

* resolves #2: added compilable kernel recipe

This kernel can be ran using qemu `runqemu x86-64` after baking `core-image-minimal`

The issue before was complaining that nothing provided `vmlinux`. The route cause of this issue was the `SRCPV` was not properly set. This was fixed by:
- Adding recipe version in recipe name
- Updating `LINUX_VERSION` variable to accurately represent the kernel version
- Modified `SRC_URI` names to `machine` for Linux source, and `meta` for kernel meta